+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage/config
 
+## 1.0.2-next.0
+
+### Patch Changes
+
+- 7d47def9c4: Removed dependency on `@types/jest`.
+
 ## 1.0.1
 
 ### Patch Changes