+++ plugins/scaffolder-common/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-scaffolder-common
 
+## 1.2.0-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.1.1-next.0
+
 ## 1.2.0-next.0
 
 ### Minor Changes